// tshe-thau-loo/src/app/models/activity/event/event-create.model.ts
export interface IEventCreate {
  title: string;
  content: string;
  declaration: string;
  venue: string;
  registrationStartDate: Date;
  registrationStartTime: Date;
  registrationEndDate: Date;
  registrationEndTime: Date;
  startDate: Date;
  startTime: Date;
  endDate: Date;
  endTime: Date;
  limitNumberOfPeople: number;
  enableVerify: boolean;
  enableIdentityConfirmed: boolean;
}